Moved create_test_df_source() back into factory.
