+++ tests/test_transforms.py
@@ -34,7 +34,6 @@ from zipline.transforms import Returns
 import zipline.utils.factory as factory
 
 from zipline.test_algorithms import BatchTransformAlgorithm
-from zipline.sources import DataFrameSource
 
 
 def to_dt(msg):
@@ -310,21 +309,10 @@ class TestFinanceTransforms(TestCase):
 ############################################################
 # Test BatchTransform
 
-def create_test_df_source():
-    start = pd.datetime(1990, 1, 3, 0, 0, 0, 0, pytz.utc)
-    end = pd.datetime(1990, 1, 8, 0, 0, 0, 0, pytz.utc)
-    index = pd.DatetimeIndex(start=start, end=end, freq=pd.datetools.day)
-    x = np.arange(2., len(index) * 2 + 2).reshape((-1, 2))
-
-    df = pd.DataFrame(x, index=index, columns=[0, 1])
-
-    return DataFrameSource(df), df
-
-
 class TestBatchTransform(TestCase):
     def setUp(self):
         setup_logger(self)
-        self.source, self.df = create_test_df_source()
+        self.source, self.df = factory.create_test_df_source()
 
     def test_event_window(self):
         algo = BatchTransformAlgorithm()
